refactor(tc): keep annotation rendering in tests
